refactor: support angular version 5
BREAKING CHANGE:
Angular version 5 comes with a breaking change which results in an internal breaking change which remove API options

Bindings are no longer supported when opening a modal.
Providing bindings was, from day one, a shortcut to a manual operation.
If you used bindings before, instead provide an injector with the bindings inside.

Most of the users should not feel this change as usage of bindings should be minimal.

If you wrote a plugin you should feel this, follow the change set for one
of the native plugins as an conversion example.