refactor: dont return a promise when opening a dialog (#383)
BREAKING CHANGE:
Calling `modal.open()` returned a Promise of `DialogRef`. The async operation is not required and exists due to legacy angular implementation of dynamic components.
`modal.open()` now returns the `DialogRef` instance and NOT the `Promise`.

You need to refactor your codebase to accomodate this change, this is a big change
and it is required to remove the complexity working with `DialogRef`.

Plugin authoers: `Maybe` type does not exists anymore.
If you retuned the dialog instance from your `Modal` implementations there is not much to refactor other then types here and there.

If you return `Promise` of `DialogRef` you will have to refactor your code.